+++ api/Cargo.toml
@@ -3,6 +3,10 @@ name = "api"
 path = "src/main.rs"
 
 [[bin]]
+name = "main"
+path = "bin/v1.rs"
+
+[[bin]]
 name = "generate-gql-schema"
 path = "src/bin/generate_gql_schema.rs"
