+++ api/src/v1/error.rs
