+++ api/src/v1/mod.rs
@@ -1,0 +1,6 @@
+pub mod models;
+pub mod error;
+pub mod shutdown;
+pub mod state;
+pub mod sink;
+pub mod routes;
