+++ api/src/v1/shutdown.rs
