--- ptolemy-api/Cargo.toml
@@ -1,103 +1,0 @@
-[package]
-name = "ptolemy-api"
-version = "0.1.0"
-edition = "2024"
-
-[[bin]]
-name = "run"
-path = "src/bin/run.rs"
-
-[dependencies]
-http = "1.2.0"
-r2d2 = "0.8.10"
-tokio-stream = "0.1.17"
-tonic-web = "0.12.3"
-
-[dependencies.axum]
-workspace = true
-
-[dependencies.bb8]
-workspace = true
-
-[dependencies.chrono]
-workspace = true
-
-[dependencies.diesel]
-workspace = true
-
-[dependencies.diesel-async]
-workspace = true
-
-[dependencies.hyper]
-workspace = true
-
-[dependencies.serde]
-workspace = true
-
-[dependencies.serde_json]
-workspace = true
-
-[dependencies.serde_repr]
-workspace = true
-
-[dependencies.time]
-workspace = true
-
-[dependencies.tokio]
-workspace = true
-
-[dependencies.uuid]
-workspace = true
-
-[dependencies.tonic]
-workspace = true
-
-[dependencies.prost]
-workspace = true
-
-[dependencies.tonic-types]
-workspace = true
-
-[dependencies.prost-types]
-workspace = true
-
-[dependencies.tracing]
-workspace = true
-
-[dependencies.tracing-subscriber]
-workspace = true
-
-[dependencies.ring]
-workspace = true
-
-[dependencies.base64]
-workspace = true
-
-[dependencies.tower-http]
-workspace = true
-features = [ "cors",]
-
-[dependencies.argon2]
-workspace = true
-
-[dependencies.tower]
-workspace = true
-
-[dependencies.jsonwebtoken]
-workspace = true
-
-[dependencies.ipnet]
-workspace = true
-
-[dependencies.heck]
-workspace = true
-
-[dependencies.pin-project-lite]
-workspace = true
-
-[dependencies.diesel_migrations]
-version = "2.2.0"
-features = [ "postgres",]
-
-[dependencies.ptolemy]
-path = "../ptolemy"
