--- ptolemy-api/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod error;
-pub mod middleware;
-pub mod models;
-pub mod routes;
-pub mod shutdown;
-pub mod sink;
-pub mod state;
